chore(MediaSettings): move components to corresponding folders
